+++ lib/main.dart
@@ -7,7 +7,7 @@ import 'models/app_model.dart';
 import 'models/planets/celestial_body.dart';
 import 'screens/screen_about.dart';
 import 'screens/screen_home.dart';
-import 'screens/settings.dart';
+import 'screens/screen_settings.dart';
 import 'screens/tabs/iss/screen_iss.dart';
 import 'screens/tabs/news/screen_news.dart';
 import 'screens/tabs/planets/add_edit_planet.dart';
@@ -24,32 +24,33 @@ void main() async {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return new ScopedModel<AppModel>(
-        model: model,
-        child: new ScopedModelDescendant<AppModel>(
-          builder: (context, child, model) => MaterialApp(
-                title: 'Space Curiosity',
-                theme: model.themeData,
-                home: HomeScreen(),
-                debugShowCheckedModeBanner: false,
-                routes: <String, WidgetBuilder>{
-                  '/home': (_) => HomeScreen(),
-                  '/spacex': (_) => SpacexScreen(),
-                  '/news': (_) => NewsScreen(),
-                  '/planets': (_) => SolarSystemScreen(),
-                  '/iss': (_) => IssScreen(),
-                  '/info': (_) => AboutScreen(),
-                  '/about': (_) => AboutScreen(),
-                  AddEditPlanetPage.routeName: (_) =>
-                      AddEditPlanetPage(null, type: BodyType.planet),
-                  '/settings': (_) => SettingsPage(),
-                },
-                localizationsDelegates: [
-                  FlutterI18nDelegate(false, 'en'),
-                  GlobalMaterialLocalizations.delegate,
-                  GlobalWidgetsLocalizations.delegate
-                ],
-              ),
-        ));
+    return ScopedModel<AppModel>(
+      model: model,
+      child: ScopedModelDescendant<AppModel>(
+        builder: (context, child, model) => MaterialApp(
+              title: 'Space Curiosity',
+              theme: model.themeData,
+              home: HomeScreen(),
+              debugShowCheckedModeBanner: false,
+              routes: <String, WidgetBuilder>{
+                '/home': (_) => HomeScreen(),
+                '/spacex': (_) => SpacexScreen(),
+                '/news': (_) => NewsScreen(),
+                '/planets': (_) => SolarSystemScreen(),
+                '/iss': (_) => IssScreen(),
+                '/info': (_) => AboutScreen(),
+                '/about': (_) => AboutScreen(),
+                AddEditPlanetPage.routeName: (_) =>
+                    AddEditPlanetPage(null, type: BodyType.planet),
+                '/settings': (_) => SettingsScreen(),
+              },
+              localizationsDelegates: [
+                FlutterI18nDelegate(false, 'en'),
+                GlobalMaterialLocalizations.delegate,
+                GlobalWidgetsLocalizations.delegate
+              ],
+            ),
+      ),
+    );
   }
 }
